requirements.txt
Flask==2.3.4
SQLAlchemy==2.0.19
Flask-Migrate==4.0.4
PyMySQL==1.1.0
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.0
Flask-Limiter==2.7.0
python-dotenv==1.0.0
alembic==1.11.1
argon2-cffi==21.3.0
gunicorn==21.2.0
pytest==7.4.2
psutil==5.9.5
marshmallow==3.20.0
marshmallow-sqlalchemy==0.28.0
qrcode==7.4.2
Pillow==10.0.0
